filter out some mailcap entries on start
Now the only entry that is filtered out on load is */*.
Because sometimes there is */* entry in a mailcap file that
acts as catch-all entry. It's more useful to show these files
as files with unknown format. This helps to detect these files
quickly.

Add -F option to disable this behaviour.